refactor code to put AuthContext in its own file
