+++ src/App.tsx
@@ -1,74 +1,15 @@
 import { BrowserRouter, Route } from 'react-router-dom';
-import { createContext, useState, useEffect } from 'react';
 import { Home } from './pages/Home';
 import { NewRoom } from './pages/NewRoom';
-import { auth, firebase } from './services/firebase';
-
-type User = {
-  id: string;
-  name: string;
-  avatar: string;
-}
-
-type AuthContextType = {
-  user: User | undefined;
-  signInWithGoogle: () => Promise<void>;
-}
-
-
-export const AuthContext = createContext({} as AuthContextType); 
+import { AuthContextProvider } from './contexts/AuthContext';
 
 function App() {
-  const [user, setUser] = useState<User>();
-
-  useEffect(() => {
-    const unsubscribe = auth.onAuthStateChanged(user => {
-      if (user) {
-        const { displayName, photoURL, uid } = user;
-
-        if (!displayName || !photoURL) {
-          throw new Error('Missing information from Google Account.')
-        }
-  
-        setUser({
-          id: uid,
-          name: displayName,
-          avatar: photoURL,
-        })
-      }
-    });
-
-    return () => {
-      unsubscribe();
-    }
-  }, [])
-
-  const signInWithGoogle = async () => {
-    const provider = new firebase.auth.GoogleAuthProvider();
-
-    const result = await auth.signInWithPopup(provider);
-
-    if (result.user) {
-      const { displayName, photoURL, uid } = result.user;
-
-      if (!displayName || !photoURL) {
-        throw new Error('Missing information from Google Account.')
-      }
-
-      setUser({
-        id: uid,
-        name: displayName,
-        avatar: photoURL,
-      })
-    }
-  }
-
   return (
     <BrowserRouter>
-      <AuthContext.Provider value={{ user, signInWithGoogle }}>
+      <AuthContextProvider>
         <Route path="/" exact component={Home} />
         <Route path="/rooms/new" component={NewRoom} />
-      </AuthContext.Provider>
+      </AuthContextProvider>
     </BrowserRouter>
   );
 }
