+++ src/pages/NewRoom.tsx
@@ -4,7 +4,7 @@ import { Link } from 'react-router-dom';
 import { Button } from '../components/Button';
 import '../assets/styles/auth.scss';
 import { useContext } from 'react';
-import { AuthContext } from '../App';
+import { AuthContext } from '../contexts/AuthContext';
 
 export const NewRoom = () => {
   const { user } = useContext(AuthContext)
@@ -18,7 +18,6 @@ export const NewRoom = () => {
     <main>
       <div className="main-content">
         <img src={logoImg} alt="Letmeask logo" />
-        <h1>{user?.name}</h1>
         <h2>Create a new room</h2>
         <form>
           <input
